pyproject.toml
[project]
name = "django-vanya"
version = "0.1.0"
description = "Учебный Django-проект для подготовки к техническому собеседованию"
readme = "README.md"
requires-python = ">=3.12.1"
dependencies = [
    "django>=6.0.4",
    "djangorestframework>=3.16.0",
    "gunicorn>=23.0.0",
    "whitenoise>=6.9.0",
]
